refactor: FunctionUtil에 확장함수 통합
