--- presentation/src/main/java/com/shypolarbear/presentation/ui/signup/custom/CustomFunctions.kt
@@ -1,27 +1,0 @@
-package com.shypolarbear.presentation.ui.signup.custom
-
-import android.content.Context
-import android.view.inputmethod.InputMethodManager
-import android.widget.TextView
-import androidx.core.content.ContextCompat
-import androidx.fragment.app.Fragment
-
-fun TextView.setTextColorById(context: Context, colorId: Int) {
-    this.setTextColor(
-        ContextCompat.getColor(
-            context,
-            colorId
-        )
-    )
-}
-
-fun Fragment.hideKeyboard() {
-    if (activity != null && requireActivity().currentFocus != null) {
-        val inputManager =
-            requireActivity().getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-        inputManager.hideSoftInputFromWindow(
-            requireActivity().currentFocus!!.windowToken,
-            InputMethodManager.HIDE_NOT_ALWAYS
-        )
-    }
-}
